[SHELL32_APITEST] ShellHook testcase Fix compiler error (addendum to #3289)
addendum to 0.4.15-dev-1123-g 0089291751837d977431c85b6d8d6b1232fbb67b

Fix a compiler error on RosBE2.1.6 GCC4.7.2
that I got when I merged this test back to 0.4.14RC:

C:/0414rls/reactos/modules/rostests/apitests/shell32/ShellHook.cpp: In function
'LRESULT WindowProc(HWND, UINT, WPARAM, LPARAM)':
C:/0414rls/reactos/modules/rostests/apitests/shell32/ShellHook.cpp:339:26: error
: comparison between signed and unsigned integer expressions [-Werror=sign-compa
re]
cc1plus.exe: all warnings being treated as errors
ninja: build stopped: subcommand failed.

+++ modules/rostests/apitests/shell32/ShellHook.cpp
@@ -336,7 +336,7 @@ WindowProc(HWND hwnd, UINT uMsg, WPARAM wParam, LPARAM lParam)
                 INT i = (INT)wParam - ID_BURNING;
                 DoTestEntryPart2(&s_entries[i]);
                 ++i;
-                if (i == s_num_entries)
+                if (i == (INT)s_num_entries)
                 {
                     PostQuitMessage(0);
                     break;
